integer argument for progresssheet maxvalue

+++ SKMainWindowController.h
@@ -312,7 +312,7 @@ enum {
 - (void)updateLeftStatus;
 - (void)updateRightStatus;
 
-- (void)beginProgressSheetWithMessage:(NSString *)message maxValue:(double)maxValue;
+- (void)beginProgressSheetWithMessage:(NSString *)message maxValue:(NSUInteger)maxValue;
 - (void)incrementProgressSheet;
 - (void)dismissProgressSheet;
